+++ src/lib.rs
@@ -6,4 +6,5 @@ pub mod domain;
 pub mod events;
 pub mod grpc;
 pub mod infra;
+pub mod init_common;
 pub mod role;